final change

+++ speak-back/speak-elastic/src/main/java/com/speak/speakelastic/service/KafkaListeners.java
@@ -3,6 +3,8 @@ package com.speak.speakelastic.service;
 import com.speak.speakelastic.documents.GlobalObject;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.kafka.annotation.KafkaListener;
+import org.springframework.kafka.support.KafkaHeaders;
+import org.springframework.messaging.handler.annotation.Header;
 import org.springframework.stereotype.Component;
 
 @Component
@@ -26,25 +28,16 @@ public class KafkaListeners {
             groupId = "myGroup"
 
     )
-    public void postListener(GlobalObject object) {
+    public void postListener(GlobalObject object,@Header("CUSTOM_HEADER") String header) {
         System.out.println(object.toString());
-        this.gObjectService.save(object);
+        System.out.println(header);
+        if(header.contains("post")){
+        this.gObjectService.save(object);}
+        else  if(header.contains("delete")){
+            this.gObjectService.delete(object.getId());
+        }
 
     }
 
-    /**
-     * This function listens for any message in the delete post
-     * at which case it delets the message from the elasticsearch database
-     *
-     * @param object
-     */
-    @KafkaListener(
-            topics = "delete",
-            groupId = "deleteID"
 
-    )
-    public void deleteListener(GlobalObject object) {
-        this.gObjectService.delete(object.getId());
-
-    }
 }
